# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Skillset API",
    "version": "0.1.0",
    "description": "A self-hosted Registry for Agent Skills. This is the contract the CLI depends on; the web interface uses the same routes and has no private data path of its own.\n\n**Conventions.** Every JSON key on the wire is snake_case, which means wire keys line up with Postgres column names. TypeScript types, interfaces, and classes are camelCase, so conversion happens once at the boundary in the shared module rather than ad hoc per call site.\n\n**Artifact bytes never pass through this API.** Publishing returns a presigned upload target that the caller writes to directly, and reading an Artifact redirects to a short-lived presigned location. The API therefore validates the metadata a caller claims about a Skill, never the Artifact itself (ADR-0001)."
  },
  "servers": [{ "url": "/api" }],
  "security": [{ "cookieAuth": [] }, { "bearerAuth": [] }],
...
  "paths": {
...
    "/identity-providers": {
      "get": {
        "tags": ["identity-providers"],
        "summary": "List every configured Identity Provider",
        "description": "Requires at least the admin role. Includes disabled Providers, and never includes a client secret.",
        "responses": {
          "200": {
            "description": "Every configured Identity Provider",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["items"],
                  "properties": {
                    "items": { "type": "array", "items": { "$ref": "#/components/schemas/IdentityProvider" } }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthenticated" },
          "403": { "$ref": "#/components/responses/AdminOnly" }
        }
      },
      "post": {
        "tags": ["identity-providers"],
        "summary": "Configure an Identity Provider",
        "description": "Requires at least the admin role. At most one Provider per kind (ADR-0017). A Provider may be created enabled, but only with a permitted organisation set — with just-in-time provisioning that gate is the only control on who gets an account (ADR-0015).\n\nThere is no issuer URL to supply: Better Auth knows each kind's endpoints (ADR-0016), so what an Admin configures is the credential pair and the organisation to admit.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["kind", "display_name", "client_id", "client_secret"],
                "properties": {
                  "kind": {
                    "type": "string",
                    "enum": ["google", "microsoft", "github"],
                    "description": "The Provider's identity, and immutable after creation (ADR-0017)."
                  },
                  "display_name": { "type": "string", "minLength": 1 },
                  "client_id": { "type": "string", "minLength": 1 },
                  "client_secret": {
                    "type": "string",
                    "minLength": 1,
                    "description": "Stored as given, and never returned by any response."
                  },
                  "permitted_organisations": { "type": "array", "items": { "type": "string", "minLength": 1 }, "maxItems": 50, "description": "Organisations this Provider admits — Workspace domains, Entra tenant ids, or GitHub organisation logins by kind. A login is admitted if it matches any one of them, case-insensitively. An empty array means no organisation check runs at all and everyone the provider authenticates gets a reader account (ADR-0021)." },
                  "enabled": { "type": "boolean", "default": false }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The configured Identity Provider",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/IdentityProvider" } } }
          },
          "400": { "$ref": "#/components/responses/ValidationFailed" },
          "401": { "$ref": "#/components/responses/Unauthenticated" },
          "403": { "$ref": "#/components/responses/AdminOnly" },
          "409": {
            "description": "An Identity Provider of that kind is already configured (`kind_taken`)",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          }
        }
      }
    },
...
    "/integrations": {
      "get": {
        "tags": ["integrations"],
        "summary": "List every configured Integration",
        "description": "Requires at least the admin role.\n\nAn Integration's existence is the only switch Importing has: it is available for exactly those Git Providers with a row, and there is deliberately no `import_enabled` flag above it (ADR-0024). An operator who wants this Registry to hold no repository credentials for anybody creates no Integration.\n\nThis is not an Identity Provider and is configured separately. Signing in with GitHub and Importing from GitHub are two registrations that share nothing but a vendor, so either may exist without the other.\n\nNever includes a client secret. There is no unauthenticated counterpart to this route — unlike an Identity Provider, an Integration draws no button on a login page.",
        "responses": {
          "200": {
            "description": "Every configured Integration",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["items"],
                  "properties": {
                    "items": { "type": "array", "items": { "$ref": "#/components/schemas/Integration" } }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthenticated" },
          "403": { "$ref": "#/components/responses/AdminOnly" }
        }
      },
      "post": {
        "tags": ["integrations"],
        "summary": "Register a Git Provider",
        "description": "Requires at least the admin role.\n\nAn Integration's existence is the only switch Importing has: it is available for exactly those Git Providers with at least one row, and there is deliberately no `import_enabled` flag above it (ADR-0024). An operator who wants this Registry to hold no repository credentials for anybody creates no Integration.\n\nThis is not an Identity Provider and is configured separately. Signing in with GitHub and Importing from GitHub are two registrations that share nothing but a vendor, so either may exist without the other.\n\nMore than one Integration may be registered for the same Git Provider — e.g. two different GitHub Apps — and a writer chooses which to connect through (ADR-0025). `provider` must name a Git Provider this Registry reads from; that list is `GIT_PROVIDERS` in the shared package, and a provider is added there rather than by migration.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["provider", "display_name", "client_id", "client_secret"],
                "properties": {
                  "provider": { "type": "string", "enum": ["github", "gitlab"], "example": "github" },
                  "display_name": { "type": "string", "maxLength": 100, "example": "GitHub" },
                  "client_id": { "type": "string", "example": "Iv1.a1b2c3d4e5f6" },
                  "client_secret": {
                    "type": "string",
                    "description": "Stored as given (ADR-0015). No response shape ever returns it."
                  },
                  "app_slug": {
                    "type": "string",
                    "nullable": true,
                    "pattern": "^[A-Za-z0-9-]+$",
                    "maxLength": 100,
                    "example": "acme-skill-registry",
                    "description": "Interpolated into the provider's installation URL, hence the narrow pattern. Omitted is the same as null: this provider has no installation step."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The created Integration",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Integration" } } }
          },
          "400": { "$ref": "#/components/responses/ValidationFailed" },
          "401": { "$ref": "#/components/responses/Unauthenticated" },
          "403": { "$ref": "#/components/responses/AdminOnly" }
        }
      }
    },
    "/integrations/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": { "type": "string", "format": "uuid" },
          "description": "The Integration's id."
        }
      ],
      "patch": {
        "tags": ["integrations"],
        "summary": "Change an Integration's configuration",
        "description": "Requires at least the admin role.\n\n`provider` is not on this request body at all: which Git Provider an app is for does not change once registered, only the app's own details do.\n\nOmitting `client_secret` leaves the stored one untouched, which is what lets an Admin edit an Integration without being shown its secret. `app_slug` distinguishes absent from null — absent leaves whatever is stored, null clears it.\n\nCredentials are read per use, so an edit takes effect on the next request with no restart (ADR-0019).\n\nThere is deliberately no delete route, matching `/identity-providers`. Removing an Integration that writers still hold Connections against would strand those credentials, and the `ON DELETE RESTRICT` on `connections.integration_id` refuses it at the database.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "display_name": { "type": "string", "maxLength": 100 },
                  "client_id": { "type": "string" },
                  "client_secret": { "type": "string" },
                  "app_slug": { "type": "string", "nullable": true, "pattern": "^[A-Za-z0-9-]+$", "maxLength": 100 }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The updated Integration",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Integration" } } }
          },
          "400": { "$ref": "#/components/responses/ValidationFailed" },
          "401": { "$ref": "#/components/responses/Unauthenticated" },
          "403": { "$ref": "#/components/responses/AdminOnly" },
          "404": {
            "description": "No Integration exists by that id",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session",
        "description": "Signed session cookie, http-only, secure, strictly same-site. Carries only the User's id; the role is resolved from the database on every request (ADR-0005). Cannot be revoked before expiry."
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "A Token, used by the CLI. Looked up by hash on every request; revocable immediately."
      }
    },
...
    "responses": {
...
      "Unauthenticated": {
        "description": "No valid session or Token. The web interface treats this as an expired session, discards cached state, and returns to login.",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
...
    "AdminOnly": {
        "description": "Requires at least the admin role (admin or the superadmin)",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
...
      "ValidationFailed": {
        "description": "The request failed a validation rule. `code` names the rule and `field` names what broke it, so the CLI and the web interface can report the same specific reason.",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
...
    },
    "schemas": {
...
      "IdentityProvider": {
        "type": "object",
        "description": "An Admin's view of a configured Identity Provider. `client_secret` is deliberately absent from this and every other response: it is written and never read back, so no response can leak it, not even to the Admin who set it.",
        "required": [
          "id",
          "kind",
          "display_name",
          "client_id",
          "permitted_organisations",
          "enabled",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "kind": { "type": "string", "enum": ["google", "microsoft", "github"] },
          "display_name": { "type": "string" },
          "client_id": { "type": "string" },
          "permitted_organisations": {
            "type": "array",
            "items": { "type": "string" },
            "description": "The organisations allowed to sign in: Workspace domains (Google `hd`), tenant ids (Entra `tid`), or GitHub organisation logins. A login is admitted if what the Provider asserts matches any one of them, case-insensitively — a claim for the first two, a live `/user/orgs` check for GitHub, which issues no ID token (ADR-0018) — and never against the email address's suffix. An empty array is a configuration rather than a blank: no check runs, and everyone the provider authenticates gets a reader account here (ADR-0021)."
          },
          "enabled": {
            "type": "boolean",
            "description": "Whether the login page offers it and logins through it are accepted. Disabling is how a Provider is taken out of service; there is no delete."
          },
          "created_at": { "type": "string", "format": "date-time" },
          "updated_at": { "type": "string", "format": "date-time" }
        }
      },
      "Integration": {
        "type": "object",
        "description": "The Registry's registration with a Git Provider, holding the credential pair a Connection is granted against (ADR-0024). `client_secret` is deliberately absent from this and every other response: it is written and never read back. More than one Integration may exist for the same `provider` (ADR-0025) — `id` is the row's identity.",
        "required": ["id", "provider", "display_name", "client_id", "app_slug", "created_at", "updated_at"],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "provider": {
            "type": "string",
            "example": "github",
            "description": "The Git Provider this app is for. Plain text rather than an enum, so registering a new provider is an insert and not a migration."
          },
          "display_name": { "type": "string", "example": "GitHub" },
          "client_id": { "type": "string", "example": "Iv1.a1b2c3d4e5f6" },
          "app_slug": {
            "type": "string",
            "nullable": true,
            "example": "acme-skill-registry",
            "description": "The app's slug at the provider, used to build its installation URL. Null for a provider with no installation step — only GitHub has one."
          },
          "created_at": { "type": "string", "format": "date-time" },
          "updated_at": { "type": "string", "format": "date-time" }
        }
      },
...
      "Error": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": {
            "type": "object",
            "required": ["code", "message"],
            "properties": {
              "code": {
                "type": "string",
                "description": "A stable machine-readable identifier for what went wrong — for validation failures, the specific rule that failed."
              },
              "message": { "type": "string" },
              "field": {
                "type": "string",
                "description": "The field or file that broke the rule, when there is one."
              }
            }
          }
        }
      }
    }
  }
}